moved new universal tests for classifiers inside teh all_classifier test
