+++ tests/classifiersFrameworks/test_keras.py
@@ -219,153 +219,6 @@ def test_defences_predict(get_default_mnist_subset, get_image_classifier_list):
 
 
 @pytest.mark.only_with_platform("keras")
-def test_loss_gradient(framework, get_default_mnist_subset, get_image_classifier_list):
-    expected_values = {
-        "expected_gradients_1": ExpectedValue(
-
-            np.asarray([
-                0.00210803,
-                0.00213919,
-                0.0052098,
-                0.00548,
-                -0.0023059,
-                0.00432076,
-                0.00274945,
-                0.,
-                0.,
-                -0.0583441,
-                -0.00616604,
-                0.05262191,
-                -0.02373985,
-                0.05273107,
-                0.10711591,
-                0.12773865,
-                0.06892889,
-                0.01337799,
-                0.1003202,
-                0.01681095,
-                -0.00028647,
-                -0.05588859,
-                0.01474165,
-                0.,
-                0.,
-                0.,
-                0.,
-                0.,
-            ]),
-            4,
-        ),
-        "expected_gradients_2": ExpectedValue(
-            np.asarray(
-                [0.0559206,
-                 0.05338925,
-                 0.0648919,
-                 0.07925165,
-                 -0.04029291,
-                 - 0.11281465,
-                 0.01850601,
-                 0.00325053,
-                 0.08163194,
-                 0.03333949,
-                 0.031766,
-                 -0.02420464,
-                 -0.07815557,
-                 -0.04698735,
-                 0.10711591,
-                 0.04086433,
-                 -0.03441072,
-                 0.01071284,
-                 - 0.04229196,
-                 - 0.01386157,
-                 0.02827487,
-                 0.,
-                 0.,
-                 0.,
-                 0.,
-                 0.,
-                 0.,
-                 0.]),
-            4,
-        ),
-    }
-
-    # expected_values = {
-    #     "expected_gradients_1": ExpectedValue(
-    #         np.asarray(
-    #             [
-    #                 0.0559206,
-    #                 0.05338925,
-    #                 0.0648919,
-    #                 0.07925165,
-    #                 -0.04029291,
-    #                 -0.11281465,
-    #                 0.01850601,
-    #                 0.00325054,
-    #                 0.08163195,
-    #                 0.03333949,
-    #                 0.031766,
-    #                 -0.02420463,
-    #                 -0.07815556,
-    #                 -0.04698735,
-    #                 0.10711591,
-    #                 0.04086434,
-    #                 -0.03441073,
-    #                 0.01071284,
-    #                 -0.04229195,
-    #                 -0.01386157,
-    #                 0.02827487,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #             ]
-    #         ),
-    #         4,
-    #     ),
-    #     "expected_gradients_2": ExpectedValue(
-    #         np.asarray(
-    #             [
-    #                 0.00210803,
-    #                 0.00213919,
-    #                 0.00520981,
-    #                 0.00548001,
-    #                 -0.0023059,
-    #                 0.00432077,
-    #                 0.00274945,
-    #                 0.0,
-    #                 0.0,
-    #                 -0.0583441,
-    #                 -0.00616604,
-    #                 0.0526219,
-    #                 -0.02373985,
-    #                 0.05273106,
-    #                 0.10711591,
-    #                 0.12773865,
-    #                 0.0689289,
-    #                 0.01337799,
-    #                 0.10032021,
-    #                 0.01681096,
-    #                 -0.00028647,
-    #                 -0.05588859,
-    #                 0.01474165,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #             ]
-    #         ),
-    #         4,
-    #     ),
-    # }
-
-    backend_test_loss_gradient(framework, get_default_mnist_subset, get_image_classifier_list, expected_values)
-
-
-@pytest.mark.only_with_platform("keras")
 def test_functional_model(get_functional_model):
     functional_model = get_functional_model
     keras_model = KerasClassifier(functional_model, clip_values=(0, 1), input_layer=1, output_layer=1)
@@ -430,548 +283,6 @@ def test_learning_phase(get_image_classifier_list):
 #     self.assertTrue(isinstance(loaded.defences[0], FeatureSqueezing))
 #
 #     os.remove(full_path)
-
-
-@pytest.mark.only_with_platform("keras")
-def test_class_gradient(framework, get_default_mnist_subset, get_image_classifier_list):
-    (x_train_mnist, y_train_mnist), (x_test_mnist, y_test_mnist) = get_default_mnist_subset
-
-    classifier, _ = get_image_classifier_list(one_classifier=True, from_logits=True)
-
-    expected_values = {
-        "expected_gradients_1_all_labels": ExpectedValue(
-            np.asarray(
-                [
-                    -0.03347399,
-                    -0.03195872,
-                    -0.02650188,
-                    0.04111874,
-                    0.08676253,
-                    0.03339913,
-                    0.06925241,
-                    0.09387045,
-                    0.15184258,
-                    -0.00684002,
-                    0.05070481,
-                    0.01409407,
-                    -0.03632583,
-                    0.00151133,
-                    0.05102589,
-                    0.00766463,
-                    -0.00898967,
-                    0.00232938,
-                    -0.00617045,
-                    -0.00201032,
-                    0.00410065,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                ]
-            ),
-            4,
-        ),
-        "expected_gradients_2_all_labels": ExpectedValue(
-            np.asarray(
-                [
-                    -0.09723657,
-                    -0.00240533,
-                    0.02445251,
-                    -0.00035474,
-                    0.04765627,
-                    0.04286841,
-                    0.07209076,
-                    0.0,
-                    0.0,
-                    -0.07938144,
-                    -0.00142567,
-                    0.02882954,
-                    -0.00049514,
-                    0.04170151,
-                    0.05102589,
-                    0.09544909,
-                    -0.04401167,
-                    -0.06158172,
-                    0.03359772,
-                    -0.00838454,
-                    0.01722163,
-                    -0.13376027,
-                    0.08206709,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                ]
-            ),
-            4,
-        ),
-        "expected_gradients_1_label5": ExpectedValue(
-            np.asarray(
-                [
-                    -0.03347399,
-                    -0.03195872,
-                    -0.02650188,
-                    0.04111874,
-                    0.08676253,
-                    0.03339913,
-                    0.06925241,
-                    0.09387045,
-                    0.15184258,
-                    -0.00684002,
-                    0.05070481,
-                    0.01409407,
-                    -0.03632583,
-                    0.00151133,
-                    0.05102589,
-                    0.00766463,
-                    -0.00898967,
-                    0.00232938,
-                    -0.00617045,
-                    -0.00201032,
-                    0.00410065,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                ]
-            ),
-            4,
-        ),
-        "expected_gradients_2_label5": ExpectedValue(
-            np.asarray(
-                [
-                    -0.09723657,
-                    -0.00240533,
-                    0.02445251,
-                    -0.00035474,
-                    0.04765627,
-                    0.04286841,
-                    0.07209076,
-                    0.0,
-                    0.0,
-                    -0.07938144,
-                    -0.00142567,
-                    0.02882954,
-                    -0.00049514,
-                    0.04170151,
-                    0.05102589,
-                    0.09544909,
-                    -0.04401167,
-                    -0.06158172,
-                    0.03359772,
-                    -0.00838454,
-                    0.01722163,
-                    -0.13376027,
-                    0.08206709,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                ]
-            ),
-            4,
-        ),
-        # THIS IS FAILING
-        "expected_gradients_1_labelArray": ExpectedValue(
-            np.asarray(
-                [
-                    0.06860766,
-                    0.065502,
-                    0.08539103,
-                    0.13868105,
-                    -0.05520725,
-                    -0.18788849,
-                    0.02264893,
-                    0.02980516,
-                    0.2226511,
-                    0.11288887,
-                    -0.00678776,
-                    0.02045561,
-                    -0.03120914,
-                    0.00642691,
-                    0.08449504,
-                    0.02848018,
-                    -0.03251382,
-                    0.00854315,
-                    -0.02354656,
-                    -0.00767687,
-                    0.01565931,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                ]
-            ),
-            4,
-        ),
-        "expected_gradients_2_labelArray": ExpectedValue(
-            np.asarray(
-                [
-                    -0.0487146,
-                    -0.0171556,
-                    -0.03161772,
-                    -0.0420007,
-                    0.03360246,
-                    -0.01864819,
-                    0.00315916,
-                    0.0,
-                    0.0,
-                    -0.07631349,
-                    -0.00374462,
-                    0.04229517,
-                    -0.01131879,
-                    0.05044588,
-                    0.08449504,
-                    0.12417868,
-                    0.07536847,
-                    0.03906382,
-                    0.09467953,
-                    0.00543209,
-                    -0.00504872,
-                    -0.03366479,
-                    -0.00385999,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                ]
-            ),
-            4,
-        ),
-    }
-
-    labels = np.random.randint(5, size=x_test_mnist.shape[0])
-    # expected_values = {
-    #     "expected_gradients_1_all_labels": ExpectedValue(
-    #         np.asarray(
-    #             [
-    #                 -1.0557447e-03,
-    #                 -1.0079544e-03,
-    #                 -7.7426434e-04,
-    #                 1.7387432e-03,
-    #                 2.1773507e-03,
-    #                 5.0880699e-05,
-    #                 1.6497371e-03,
-    #                 2.6113100e-03,
-    #                 6.0904310e-03,
-    #                 4.1080985e-04,
-    #                 2.5268078e-03,
-    #                 -3.6661502e-04,
-    #                 -3.0568996e-03,
-    #                 -1.1665225e-03,
-    #                 3.8904310e-03,
-    #                 3.1726385e-04,
-    #                 1.3203260e-03,
-    #                 -1.1720930e-04,
-    #                 -1.4315104e-03,
-    #                 -4.7676818e-04,
-    #                 9.7251288e-04,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #             ]
-    #         ),
-    #         4,
-    #     ),
-    #     "expected_gradients_2_all_labels": ExpectedValue(
-    #         np.asarray(
-    #             [
-    #                 -0.00367321,
-    #                 -0.0002892,
-    #                 0.00037825,
-    #                 -0.00053344,
-    #                 0.00192121,
-    #                 0.00112047,
-    #                 0.0023135,
-    #                 0.0,
-    #                 0.0,
-    #                 -0.00391743,
-    #                 -0.0002264,
-    #                 0.00238103,
-    #                 -0.00073711,
-    #                 0.00270405,
-    #                 0.00389043,
-    #                 0.00440818,
-    #                 -0.00412769,
-    #                 -0.00441794,
-    #                 0.00081916,
-    #                 -0.00091284,
-    #                 0.00119645,
-    #                 -0.00849089,
-    #                 0.00547925,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #             ]
-    #         ),
-    #         4,
-    #     ),
-    #     "expected_gradients_1_label5": ExpectedValue(
-    #         np.asarray(
-    #             [
-    #                 -1.0557447e-03,
-    #                 -1.0079544e-03,
-    #                 -7.7426434e-04,
-    #                 1.7387432e-03,
-    #                 2.1773507e-03,
-    #                 5.0880699e-05,
-    #                 1.6497371e-03,
-    #                 2.6113100e-03,
-    #                 6.0904310e-03,
-    #                 4.1080985e-04,
-    #                 2.5268078e-03,
-    #                 -3.6661502e-04,
-    #                 -3.0568996e-03,
-    #                 -1.1665225e-03,
-    #                 3.8904310e-03,
-    #                 3.1726385e-04,
-    #                 1.3203260e-03,
-    #                 -1.1720930e-04,
-    #                 -1.4315104e-03,
-    #                 -4.7676818e-04,
-    #                 9.7251288e-04,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #             ]
-    #         ),
-    #         4,
-    #     ),
-    #     "expected_gradients_2_label5": ExpectedValue(
-    #         np.asarray(
-    #             [
-    #                 -0.00367321,
-    #                 -0.0002892,
-    #                 0.00037825,
-    #                 -0.00053344,
-    #                 0.00192121,
-    #                 0.00112047,
-    #                 0.0023135,
-    #                 0.0,
-    #                 0.0,
-    #                 -0.00391743,
-    #                 -0.0002264,
-    #                 0.00238103,
-    #                 -0.00073711,
-    #                 0.00270405,
-    #                 0.00389043,
-    #                 0.00440818,
-    #                 -0.00412769,
-    #                 -0.00441794,
-    #                 0.00081916,
-    #                 -0.00091284,
-    #                 0.00119645,
-    #                 -0.00849089,
-    #                 0.00547925,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #             ]
-    #         ),
-    #         4,
-    #     ),
-    #     "expected_gradients_1_labelArray": ExpectedValue(
-    #         np.asarray(
-    #             [
-    #                 5.0867125e-03,
-    #                 4.8564528e-03,
-    #                 6.1040390e-03,
-    #                 8.6531248e-03,
-    #                 -6.0958797e-03,
-    #                 -1.4114540e-02,
-    #                 -7.1085989e-04,
-    #                 -5.0330814e-04,
-    #                 1.2943064e-02,
-    #                 8.2416134e-03,
-    #                 -1.9859476e-04,
-    #                 -9.8109958e-05,
-    #                 -3.8902222e-03,
-    #                 -1.2945873e-03,
-    #                 7.5137997e-03,
-    #                 1.7720886e-03,
-    #                 3.1399424e-04,
-    #                 2.3657181e-04,
-    #                 -3.0891625e-03,
-    #                 -1.0211229e-03,
-    #                 2.0828887e-03,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #                 0.0000000e00,
-    #             ]
-    #         ),
-    #         4,
-    #     ),
-    #     "expected_gradients_2_labelArray": ExpectedValue(
-    #         np.asarray(
-    #             [
-    #                 -0.00195835,
-    #                 -0.00134457,
-    #                 -0.00307221,
-    #                 -0.00340564,
-    #                 0.00175022,
-    #                 -0.00239714,
-    #                 -0.00122619,
-    #                 0.0,
-    #                 0.0,
-    #                 -0.00520899,
-    #                 -0.00046105,
-    #                 0.00414874,
-    #                 -0.00171095,
-    #                 0.00429184,
-    #                 0.0075138,
-    #                 0.00792442,
-    #                 0.0019566,
-    #                 0.00035517,
-    #                 0.00504575,
-    #                 -0.00037397,
-    #                 0.00022343,
-    #                 -0.00530034,
-    #                 0.0020528,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #                 0.0,
-    #             ]
-    #         ),
-    #         4,
-    #     ),
-    # }
-    #
-    labels = np.asarray(
-        [
-            3,
-            4,
-            4,
-            0,
-            1,
-            1,
-            1,
-            2,
-            3,
-            4,
-            4,
-            2,
-            2,
-            0,
-            0,
-            4,
-            0,
-            1,
-            2,
-            0,
-            3,
-            4,
-            2,
-            2,
-            3,
-            3,
-            0,
-            1,
-            3,
-            0,
-            3,
-            2,
-            3,
-            4,
-            1,
-            3,
-            3,
-            3,
-            2,
-            1,
-            3,
-            4,
-            2,
-            3,
-            4,
-            1,
-            4,
-            0,
-            4,
-            1,
-            1,
-            4,
-            1,
-            4,
-            0,
-            1,
-            0,
-            0,
-            4,
-            0,
-            4,
-            2,
-            3,
-            1,
-            2,
-            2,
-            4,
-            3,
-            4,
-            2,
-            2,
-            4,
-            4,
-            2,
-            1,
-            3,
-            2,
-            1,
-            4,
-            1,
-            0,
-            1,
-            2,
-            1,
-            2,
-            1,
-            2,
-            1,
-            1,
-            4,
-            1,
-            2,
-            4,
-            0,
-            4,
-            1,
-            2,
-            1,
-            1,
-            3,
-        ]
-    )
-
-    backend_test_class_gradient(framework, get_default_mnist_subset, classifier, expected_values, labels)
 
 
 @pytest.mark.only_with_platform("keras")
